// src/app/models/interface.ts
export interface Details {
    id: number;
    name: string;
    gender: string;
    edu: string;
    place: string;
    dob: string;
    age: number;
    designation: string;
    experience: number;
    tech: string;
    mailId: string;
    phone: string;
    img: string;
    dateApply: number;
    coverLetter: string;
  }
export interface Status {
    name: string;
    designation: string;
    status: string;
    id: string;
}
    